move post-handshake operations to separate module
