+++ handshake/src/lib.rs
@@ -5,94 +5,30 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
 pub mod async_std;
+mod constants;
+mod error;
+pub mod post_handshake;
 pub mod sync;
 #[macro_use]
 mod utils;
 mod version;
 
-use std::{
-    cmp,
-    fmt::{self, Display},
-    io::{Read, Write},
-};
-
 use desert::{FromBytes, ToBytes};
-use futures_util::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 use log::warn;
 use snow::{
     Builder as NoiseBuilder, HandshakeState as NoiseHandshakeState,
     TransportState as NoiseTransportState,
 };
 
-pub use crate::version::Version;
+use constants::{
+    EPHEMERAL_AND_STATIC_KEY_BYTES_LEN, EPHEMERAL_KEY_BYTES_LEN, PUBLIC_KEY_BYTES_LEN,
+    STATIC_KEY_BYTES_LEN,
+};
+
+pub use crate::{error::HandshakeError, version::Version};
 
 pub type Error = Box<dyn std::error::Error + Send + Sync>;
 pub type Result<T> = std::result::Result<T, Error>;
-
-#[derive(Debug, PartialEq)]
-/// Error that can occur during the handshake.
-pub enum HandshakeError {
-    /// The received major server version does not match that of the client.
-    IncompatibleServerVersion { received: u8, expected: u8 },
-}
-
-impl std::error::Error for HandshakeError {}
-
-impl Display for HandshakeError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            HandshakeError::IncompatibleServerVersion { received, expected } => {
-                write!(
-                    f,
-                    "Received server major version `{}` does not match client major version `{}`",
-                    received, expected
-                )
-            }
-        }
-    }
-}
-
-/// Size of the version message.
-const VERSION_BYTES_LEN: usize = 2;
-
-/// Number of bytes that will be written to the `send_buf` and `recv_buf`
-/// during the version exchange.
-const fn version_bytes_len() -> usize {
-    VERSION_BYTES_LEN
-}
-
-/// Size of the ephemeral key.
-const EPHEMERAL_KEY_BYTES_LEN: usize = 48;
-
-/// Number of bytes that will be written to the `send_buf` and `recv_buf`
-/// during the ephemeral key exchange.
-const fn ephemeral_key_bytes_len() -> usize {
-    EPHEMERAL_KEY_BYTES_LEN
-}
-
-/// Size of the ephemeral and static keys.
-const EPHEMERAL_AND_STATIC_KEY_BYTES_LEN: usize = 96;
-
-/// Number of bytes that will be written to the `send_buf` and `recv_buf`
-/// during the ephemeral and static key exchange.
-const fn ephemeral_and_static_key_bytes_len() -> usize {
-    EPHEMERAL_AND_STATIC_KEY_BYTES_LEN
-}
-
-/// Size of the static key.
-///
-/// In the context of the Cable Handshake implementation, this value is the
-/// private key of the author keypair.
-const STATIC_KEY_BYTES_LEN: usize = 64;
-
-/// Number of bytes that will be written to the `send_buf` and `recv_buf`
-/// during the static key exchange.
-const fn static_key_bytes_len() -> usize {
-    STATIC_KEY_BYTES_LEN
-}
-
-/// Size of the public key received via the static key exchange.
-const PUBLIC_KEY_BYTES_LEN: usize = 32;
 
 /// The initialization data of a handshake that exists in every state of the
 /// handshake.
@@ -351,7 +287,7 @@ impl Handshake<ClientSendEphemeralKey> {
         mut self,
         send_buf: &mut [u8],
     ) -> Result<Handshake<ClientRecvEphemeralAndStaticKey>> {
-        let mut write_buf = [0; ephemeral_key_bytes_len()];
+        let mut write_buf = [0; EPHEMERAL_KEY_BYTES_LEN];
 
         // Send the client ephemeral key to the server.
         let len = self.state.0.write_message(&[], &mut write_buf)?;
@@ -375,7 +311,7 @@ impl Handshake<ClientRecvEphemeralAndStaticKey> {
         mut self,
         recv_buf: &mut [u8],
     ) -> Result<Handshake<ClientSendStaticKey>> {
-        let mut read_buf = [0u8; ephemeral_and_static_key_bytes_len()];
+        let mut read_buf = [0u8; EPHEMERAL_AND_STATIC_KEY_BYTES_LEN];
 
         // Receive the ephemeral and static keys from the server.
         self.state.0.read_message(recv_buf, &mut read_buf)?;
@@ -403,7 +339,7 @@ impl Handshake<ClientSendStaticKey> {
         mut self,
         send_buf: &mut [u8],
     ) -> Result<Handshake<ClientInitTransportMode>> {
-        let mut write_buf = [0u8; static_key_bytes_len()];
+        let mut write_buf = [0u8; STATIC_KEY_BYTES_LEN];
 
         // Send the client static key to the server.
         let len = self.state.0.write_message(&[], &mut write_buf)?;
@@ -602,242 +538,13 @@ impl Handshake<ServerInitTransportMode> {
     }
 }
 
-impl Handshake<HandshakeComplete> {
-    /// Read an encrypted message from the receive buffer, decrypt and write it
-    /// to the message buffer - returning the byte size of the written payload.
-    fn read_noise_message(&mut self, recv_buf: &[u8], msg: &mut [u8]) -> Result<usize> {
-        let len = self.state.0.read_message(recv_buf, msg)?;
-
-        Ok(len)
-    }
-
-    /// Read an encrypted message of the given length from the receive buffer,
-    /// decrypt and return it as a byte vector.
-    ///
-    /// This method handles defragmentation of large (> 65519 byte) messages
-    /// automatically.
-    pub fn read_message(&mut self, recv_buf: &[u8], msg_len: u32) -> Result<Vec<u8>> {
-        // Initialise the byte indexes.
-        let mut bytes_read = 0;
-        let mut bytes_remaining = msg_len;
-
-        // Buffer to which the decrypted Noise transport message bytes for each
-        // segment will be written.
-        let mut segment_buf = vec![0u8; 65535];
-
-        // Vector to which all decrypted message segments will be appended.
-        let mut msg = Vec::new();
-
-        while bytes_remaining > 0 {
-            // Determine the byte length of the segment to be read.
-            let segment_len = cmp::min(65535, bytes_remaining);
-
-            let len = self.read_noise_message(
-                &recv_buf[bytes_read..bytes_read + segment_len as usize],
-                &mut segment_buf,
-            )?;
-
-            // Append the decrypted message segment to the complete message.
-            msg.extend_from_slice(&segment_buf[..len]);
-
-            // Update the byte indexes.
-            bytes_remaining -= segment_len;
-            bytes_read += segment_len as usize;
-        }
-
-        Ok(msg)
-    }
-
-    /// Read an encrypted message from the given stream and return it as a
-    /// byte vector.
-    ///
-    /// This method essentially reads two messages from the stream: the
-    /// unencrypted message length specifier (4 bytes) followed by the
-    /// encrypted message payload.
-    ///
-    /// An `Ok` return value containing a `Vec` of zero length and capacity
-    /// indicates receipt of an end-of-stream marker.
-    pub fn read_message_from_stream<T: Read + Write>(&mut self, stream: &mut T) -> Result<Vec<u8>> {
-        // Read four bytes describing the length of the incoming message.
-        let mut len_buf = [0; 4];
-        stream.read_exact(&mut len_buf)?;
-        let msg_len = u32::from_le_bytes(len_buf);
-
-        if msg_len == 0 {
-            // Return a 0 capacity vector to indicate end-of-stream.
-            //
-            // This does not result in an allocation.
-            return Ok(Vec::with_capacity(0));
-        }
-
-        // Read the encrypted bytes of the incoming message.
-        let mut recv_buf = vec![0u8; msg_len as usize];
-        stream.read_exact(&mut recv_buf[..])?;
-
-        // Decrypt and return the entire message.
-        let msg = self.read_message(&recv_buf, msg_len)?;
-
-        Ok(msg)
-    }
-
-    /// Read an encrypted message from the given asynchronous stream and
-    /// return it as a byte vector.
-    ///
-    /// This method essentially reads two messages from the stream: the
-    /// unencrypted message length specifier (4 bytes) followed by the
-    /// encrypted message payload.
-    ///
-    /// An `Ok` return value containing a `Vec` of zero length and capacity
-    /// indicates receipt of an end-of-stream marker.
-    pub async fn read_message_from_async_stream<T: AsyncRead + AsyncWrite + Unpin>(
-        &mut self,
-        stream: &mut T,
-    ) -> Result<Vec<u8>> {
-        // Read four bytes describing the length of the incoming message.
-        let mut len_buf = [0; 4];
-        stream.read_exact(&mut len_buf).await?;
-        let msg_len = u32::from_le_bytes(len_buf);
-
-        if msg_len == 0 {
-            // Return a 0 capacity vector to indicate end-of-stream.
-            //
-            // This does not result in an allocation.
-            return Ok(Vec::with_capacity(0));
-        }
-
-        // Read the encrypted bytes of the incoming message.
-        let mut recv_buf = vec![0u8; msg_len as usize];
-        stream.read_exact(&mut recv_buf[..]).await?;
-
-        // Decrypt and return the entire message.
-        let msg = self.read_message(&recv_buf, msg_len)?;
-
-        Ok(msg)
-    }
-
-    /// Encrypt and write a message to the send buffer, returning the byte size
-    /// of the written payload.
-    fn write_noise_message(&mut self, msg: &[u8], send_buf: &mut [u8]) -> Result<usize> {
-        let len = self.state.0.write_message(msg, send_buf)?;
-
-        Ok(len)
-    }
-
-    /// Encrypt and write a message to the send buffer, returning the total
-    /// number of written bytes.
-    ///
-    /// This method handles message fragmentation in the case that the message
-    /// byte length is greater that 65519 bytes.
-    pub fn write_message(&mut self, msg: &[u8]) -> Result<(usize, Vec<u8>)> {
-        // The length of the authentication tag included with each encrypted
-        // Noise transport message.
-        const AUTH_TAG_LEN: usize = 16;
-
-        // Initialise the byte indexes.
-        let mut bytes_written = 0;
-        let mut encrypted_bytes_written = 0;
-
-        // Buffer to which the encrypted Noise transport message bytes for each
-        // segment will be written.
-        let mut segment_buf = [0u8; 65535];
-
-        // Vector to which all encrypted message segments will be appended.
-        let mut encrypted_msg = Vec::new();
-
-        let msg_len = msg.len();
-
-        // Encrypt and append one or more message segments until the entire
-        // message has been written.
-        while bytes_written < msg_len {
-            // Determine the byte length of the segment to be written.
-            let segment_len = cmp::min(65535 - AUTH_TAG_LEN, msg_len - bytes_written);
-            // Index the bytes of the segment to be written.
-            let bytes = &msg[bytes_written..bytes_written + segment_len];
-
-            // Write the message segment.
-            let len = self.write_noise_message(bytes, &mut segment_buf)?;
-
-            // Append the encrypted message segment to the complete message.
-            encrypted_msg.extend_from_slice(&segment_buf[..len]);
-
-            // Update the byte indexes.
-            bytes_written += segment_len;
-            encrypted_bytes_written += len;
-        }
-
-        Ok((encrypted_bytes_written, encrypted_msg))
-    }
-
-    /// Encrypt and write a message to the stream, returning the byte size
-    /// of the written payload.
-    ///
-    /// This method essentially writes at least two messages to the stream:
-    /// the unencrypted message length specifier (4 bytes) followed by the
-    /// encrypted message payload. The encrypted message payload may be split
-    /// into several fragments depending on total payload size.
-    pub fn write_message_to_stream<T: Read + Write>(
-        &mut self,
-        stream: &mut T,
-        msg: &[u8],
-    ) -> Result<usize> {
-        let (bytes_written, encrypted_msg) = self.write_message(msg)?;
-        let encrypted_msg_len = &bytes_written.to_le_bytes()[..4];
-
-        stream.write_all(encrypted_msg_len)?;
-        stream.write_all(&encrypted_msg)?;
-
-        Ok(bytes_written)
-    }
-
-    /// Write an end-of-stream marker (zero-length message) to the stream.
-    pub fn write_eos_marker_to_stream<T: Read + Write>(&mut self, stream: &mut T) -> Result<()> {
-        // End-of-stream marker (message with length of 0).
-        stream.write_all(&[0, 0, 0, 0])?;
-
-        Ok(())
-    }
-
-    /// Encrypt and write a message to the asynchronous stream, returning the
-    /// byte size of the written payload.
-    ///
-    /// This method essentially writes at least two messages to the stream:
-    /// the unencrypted message length specifier (4 bytes) followed by the
-    /// encrypted message payload. The encrypted message payload may be split
-    /// into several fragments depending on total payload size.
-    pub async fn write_message_to_async_stream<T: AsyncRead + AsyncWrite + Unpin>(
-        &mut self,
-        stream: &mut T,
-        msg: &[u8],
-    ) -> Result<usize> {
-        let (bytes_written, encrypted_msg) = self.write_message(msg)?;
-        let encrypted_msg_len = &bytes_written.to_le_bytes()[..4];
-
-        stream.write_all(encrypted_msg_len).await?;
-        stream.write_all(&encrypted_msg).await?;
-
-        Ok(bytes_written)
-    }
-
-    /// Write an end-of-stream marker (zero-length message) to the asynchronous
-    /// stream.
-    pub async fn write_eos_marker_to_async_stream<T: AsyncRead + AsyncWrite + Unpin>(
-        &mut self,
-        stream: &mut T,
-    ) -> Result<()> {
-        // End-of-stream marker (message with length of 0).
-        stream.write_all(&[0, 0, 0, 0]).await?;
-
-        Ok(())
-    }
-
-    /// Return the public key of the remote peer.
-    pub fn get_remote_public_key(&self) -> Option<[u8; PUBLIC_KEY_BYTES_LEN]> {
-        self.base.remote_public_key
-    }
-}
-
 #[cfg(test)]
 mod tests {
+    use constants::{
+        EPHEMERAL_AND_STATIC_KEY_BYTES_LEN, EPHEMERAL_KEY_BYTES_LEN, STATIC_KEY_BYTES_LEN,
+        VERSION_BYTES_LEN,
+    };
+
     use super::*;
 
     fn init_handshakers(
@@ -881,16 +588,16 @@ mod tests {
 
         let mut buf = [0; 8];
 
-        let mut client_buf = &mut buf[..version_bytes_len()];
+        let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
         let hs_client = hs_client.send_client_version(&mut client_buf)?;
 
-        let mut server_buf = &mut buf[..version_bytes_len()];
+        let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
         let hs_server = hs_server.recv_client_version(&mut server_buf)?;
 
-        let mut server_buf = &mut buf[..version_bytes_len()];
+        let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
         hs_server.send_server_version(&mut server_buf)?;
 
-        let mut client_buf = &mut buf[..version_bytes_len()];
+        let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
         hs_client.recv_server_version(&mut client_buf)?;
 
         Ok(())
@@ -902,16 +609,16 @@ mod tests {
 
         let mut buf = [0; 8];
 
-        let mut client_buf = &mut buf[..version_bytes_len()];
+        let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
         let hs_client = hs_client.send_client_version(&mut client_buf)?;
 
-        let mut server_buf = &mut buf[..version_bytes_len()];
+        let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
         let hs_server = hs_server.recv_client_version(&mut server_buf)?;
 
-        let mut server_buf = &mut buf[..version_bytes_len()];
+        let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
         let _hs_server = hs_server.send_server_version(&mut server_buf)?;
 
-        let mut client_buf = &mut buf[..version_bytes_len()];
+        let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
         let hs_client = hs_client.recv_server_version(&mut client_buf);
 
         assert!(hs_client.is_err());
@@ -938,18 +645,18 @@ mod tests {
 
         // Send and receive client version.
         let (hs_client, hs_server) = {
-            let mut client_buf = &mut buf[..version_bytes_len()];
+            let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
             let hs_client = hs_client.send_client_version(&mut client_buf)?;
-            let mut server_buf = &mut buf[..version_bytes_len()];
+            let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
             let hs_server = hs_server.recv_client_version(&mut server_buf)?;
             (hs_client, hs_server)
         };
 
         // Send and receive server version.
         let (hs_client, hs_server) = {
-            let mut server_buf = &mut buf[..version_bytes_len()];
+            let mut server_buf = &mut buf[..VERSION_BYTES_LEN];
             let hs_server = hs_server.send_server_version(&mut server_buf)?;
-            let mut client_buf = &mut buf[..version_bytes_len()];
+            let mut client_buf = &mut buf[..VERSION_BYTES_LEN];
             let hs_client = hs_client.recv_server_version(&mut client_buf)?;
             (hs_client, hs_server)
         };
@@ -964,7 +671,7 @@ mod tests {
         // Send and receive client ephemeral key.
         let (hs_client, hs_server) = {
             let hs_client = hs_client.send_client_ephemeral_key(&mut buf)?;
-            let mut server_buf = &mut buf[..ephemeral_key_bytes_len()];
+            let mut server_buf = &mut buf[..EPHEMERAL_KEY_BYTES_LEN];
             let hs_server = hs_server.recv_client_ephemeral_key(&mut server_buf)?;
             (hs_client, hs_server)
         };
@@ -972,7 +679,7 @@ mod tests {
         // Send and receive server ephemeral and static keys.
         let (hs_client, hs_server) = {
             let hs_server = hs_server.send_server_ephemeral_and_static_key(&mut buf)?;
-            let mut client_buf = &mut buf[..ephemeral_and_static_key_bytes_len()];
+            let mut client_buf = &mut buf[..EPHEMERAL_AND_STATIC_KEY_BYTES_LEN];
             let hs_client = hs_client.recv_server_ephemeral_and_static_key(&mut client_buf)?;
             (hs_client, hs_server)
         };
@@ -980,7 +687,7 @@ mod tests {
         // Send and receive client static key.
         let (hs_client, hs_server) = {
             let hs_client = hs_client.send_client_static_key(&mut buf)?;
-            let mut server_buf = &mut buf[..static_key_bytes_len()];
+            let mut server_buf = &mut buf[..STATIC_KEY_BYTES_LEN];
             let hs_server = hs_server.recv_client_static_key(&mut server_buf)?;
             (hs_client, hs_server)
         };
